Begin adding hash_and_compare() code
This will replace the existing system where full file hashes are
done and then byte-for-byte comparisons are done with a unified
pair of hash-and-compares. For the standard use case, this means
that the hashes and byte-for-byte comparisons will now happen at
the same time, cutting total disk I/O by nearly 1/2 and speeding
up jdupes considerably.