Genericize BTRFS dedup to FIDEDUPERANGE
Linux kernel since 4.5 has renamed the identifier to FIDEDUPERANGE and
moved it out of the btrfs-specific space. We should expect to see more
filesystems support it.

Dedup on Windows and MacOS will be very different: although both
systems have deduplicating filesystems, there is no simple API that
does the comparison for us. Instead there is clonefile(2) on masOS
that resembles a whole-file copy (with all the permission issues that
come with it) or hardlink and FSCTL_DUPLICATE_EXTENTS_TO_FILE_EX on
Windows that needs our own block chunking and handling. The Windows
headers vaguely suggest an undocumented FSCTL_DEDUP_FILE that may or
may not do it for us.